Rider code refactoring, cleanup, reformat, warnings fixes

+++ VSDebugCoreLib/Commands/Memory/DumpMem.cs
@@ -7,11 +7,11 @@ namespace VSDebugCoreLib.Commands.Memory
 {
     internal class DumpMem : BaseCommand
     {
-        private const char tkn_force = 'f';
-        private const char tkn_append = 'a';
+        private const char TknForce = 'f';
+        private const char TknAppend = 'a';
 
         public DumpMem(VSDebugContext context)
-            : base(context, (int)PkgCmdIDList.CmdIDAbout, Resources.DumpMemString)
+            : base(context, (int) PkgCmdIDList.CmdIDAbout, Resources.DumpMemString)
         {
             CommandDescription = Resources.CmdDumpMemDesc;
 
@@ -19,27 +19,24 @@ namespace VSDebugCoreLib.Commands.Memory
                                 "\tEX: " + CommandString + " c:\\memdump.bin 0x00656789 200\n" +
                                 "\tEX: " + CommandString + " -f c:\\memdump.bin 0x00656789 200\n" +
                                 "\tFlags:\n" +
-                                "\t\t  -" + tkn_force + "   - Force file overwrite.\n" +
-                                "\t\t  -" + tkn_append + "   - Append to the file.\n" +
+                                "\t\t  -" + TknForce + "   - Force file overwrite.\n" +
+                                "\t\t  -" + TknAppend + "   - Append to the file.\n" +
                                 "\t<filename> - output filename\n" +
                                 "\t<address>  - read address, must be a hex address / pointer, can be an expression\n" +
                                 "\t<size>     - size in bytes, can be an expression\n";
 
-            CommandStatusFlag = eCommandStatus.CommandStatus_Disabled;
+            CommandStatusFlag = ECommandStatus.CommandStatusDisabled;
         }
 
-        public override eCommandStatus CommandStatus
+        public override ECommandStatus CommandStatus
         {
             get
             {
-                if (null != Context.IDE.Debugger && null != Context.IDE.Debugger.DebuggedProcesses && Context.IDE.Debugger.DebuggedProcesses.Count > 0)
-                {
-                    CommandStatusFlag = eCommandStatus.CommandStatus_Enabled;
-                }
+                if (null != Context.IDE.Debugger && null != Context.IDE.Debugger.DebuggedProcesses &&
+                    Context.IDE.Debugger.DebuggedProcesses.Count > 0)
+                    CommandStatusFlag = ECommandStatus.CommandStatusEnabled;
                 else
-                {
-                    CommandStatusFlag = eCommandStatus.CommandStatus_Disabled;
-                }
+                    CommandStatusFlag = ECommandStatus.CommandStatusDisabled;
 
                 return CommandStatusFlag;
             }
@@ -49,8 +46,8 @@ namespace VSDebugCoreLib.Commands.Memory
         {
             base.Execute(text);
 
-            char[] sp = new char[] { ' ', '\t' };
-            string[] argv = text.Split(sp, 2, StringSplitOptions.RemoveEmptyEntries);
+            char[] sp = {' ', '\t'};
+            var argv = text.Split(sp, 2, StringSplitOptions.RemoveEmptyEntries);
 
             if (argv.Length < 2)
             {
@@ -59,18 +56,18 @@ namespace VSDebugCoreLib.Commands.Memory
             }
 
             // check for optional parameters
-            string strArgParam = argv[0];
-            string strArgFile = "";
-            string strArgSource = "";
-            string strArgSize = "";
-            char chrFlag = '\0';
+            var strArgParam = argv[0];
+            var strArgFile = "";
+            var strArgSource = "";
+            var strArgSize = "";
+            var chrFlag = '\0';
 
             if ('-' == strArgParam[0])
             {
                 // max 2 characters for flags
                 if (2 == strArgParam.Length)
                 {
-                    chrFlag = Char.ToLower(strArgParam[1]);
+                    chrFlag = char.ToLower(strArgParam[1]);
                 }
                 else
                 {
@@ -79,7 +76,7 @@ namespace VSDebugCoreLib.Commands.Memory
                 }
             }
 
-            int reqNArg = ('\0' == chrFlag) ? 3 : 4;
+            var reqNArg = '\0' == chrFlag ? 3 : 4;
             argv = text.Split(sp, reqNArg, StringSplitOptions.RemoveEmptyEntries);
 
             if (argv.Length != reqNArg)
@@ -88,21 +85,18 @@ namespace VSDebugCoreLib.Commands.Memory
                 return;
             }
 
-            strArgFile = ('\0' == chrFlag) ? argv[0] : argv[1];
-            strArgSource = ('\0' == chrFlag) ? argv[1] : argv[2];
-            strArgSize = ('\0' == chrFlag) ? argv[2] : argv[3];
+            strArgFile = '\0' == chrFlag ? argv[0] : argv[1];
+            strArgSource = '\0' == chrFlag ? argv[1] : argv[2];
+            strArgSize = '\0' == chrFlag ? argv[2] : argv[3];
 
             // get file path
-            string strPath = Path.GetDirectoryName(strArgFile);
+            var strPath = Path.GetDirectoryName(strArgFile);
 
             // check if we have full path
             if (strPath != string.Empty)
             {
                 // create folder if it doesn`t exist
-                if (!Directory.Exists(strPath))
-                {
-                    Directory.CreateDirectory(strPath);
-                }
+                if (!Directory.Exists(strPath)) Directory.CreateDirectory(strPath);
             }
             // then this was meant to be in the working directory
             else
@@ -112,23 +106,20 @@ namespace VSDebugCoreLib.Commands.Memory
                 strPath = Path.GetDirectoryName(strArgFile);
 
                 // create folder if it doesn`t exist
-                if (!Directory.Exists(strPath))
-                {
-                    Directory.CreateDirectory(strPath);
-                }
+                if (!Directory.Exists(strPath)) Directory.CreateDirectory(strPath);
             }
 
             // if append or force is not specified, return error
-            if (File.Exists(strArgFile) && !(tkn_force == chrFlag || tkn_append == chrFlag))
+            if (File.Exists(strArgFile) && !(TknForce == chrFlag || TknAppend == chrFlag))
             {
                 Context.CONSOLE.Write("Output file name: " + strArgFile + " is in use!");
-                Context.CONSOLE.Write("Use -" + tkn_force + "/-" + tkn_append + " to overwrite/append");
+                Context.CONSOLE.Write("Use -" + TknForce + "/-" + TknAppend + " to overwrite/append");
                 return;
             }
 
             var varArgSource = Context.IDE.Debugger.GetExpression(strArgSource, false, 100);
             var varArgSize = Context.IDE.Debugger.GetExpression(strArgSize, false, 100);
-            int processId = Context.IDE.Debugger.CurrentProcess.ProcessID;
+            var processId = Context.IDE.Debugger.CurrentProcess.ProcessID;
 
             if (!varArgSource.IsValidValue)
             {
@@ -144,9 +135,13 @@ namespace VSDebugCoreLib.Commands.Memory
 
             long startAddress = 0;
             long dataSize = 0;
-            NumberStyles numStyleSource = NumberHelpers.IsHexNumber(varArgSource.Value) ? NumberStyles.HexNumber : NumberStyles.Integer;
-            NumberStyles numStyleSize = NumberHelpers.IsHexNumber(varArgSize.Value) ? NumberStyles.HexNumber : NumberStyles.Integer;
-            bool bRet = true;
+            var numStyleSource = NumberHelpers.IsHexNumber(varArgSource.Value)
+                ? NumberStyles.HexNumber
+                : NumberStyles.Integer;
+            var numStyleSize = NumberHelpers.IsHexNumber(varArgSize.Value)
+                ? NumberStyles.HexNumber
+                : NumberStyles.Integer;
+            var bRet = true;
 
             bRet = bRet && NumberHelpers.TryParseLong(varArgSource.Value, numStyleSource, out startAddress);
             bRet = bRet && NumberHelpers.TryParseLong(varArgSize.Value, numStyleSize, out dataSize);
@@ -157,37 +152,33 @@ namespace VSDebugCoreLib.Commands.Memory
                 return;
             }
 
-            FileMode fileMode = FileMode.Create;
+            var fileMode = FileMode.Create;
             // if force flag is specified, and the file exists, then delete it
-            if (tkn_force == chrFlag && File.Exists(strArgFile))
-            {
-                File.Delete(strArgFile);
-            }
+            if (TknForce == chrFlag && File.Exists(strArgFile)) File.Delete(strArgFile);
 
-            if (tkn_append == chrFlag && File.Exists(strArgFile))
-            {
-                fileMode = FileMode.Append;
-            }
+            if (TknAppend == chrFlag && File.Exists(strArgFile)) fileMode = FileMode.Append;
 
             if (!DebugHelpers.IsMiniDumpProcess(Context.IDE.Debugger.CurrentProcess))
             {
-                int ntdbgStatus = NativeMethods.NTDBG_OK;
-                if (NativeMethods.NTDBG_OK != (ntdbgStatus = MemoryHelpers.WriteMemoryToFile(strArgFile, processId, startAddress, dataSize, fileMode)))
+                var ntdbgStatus = NativeMethods.NtdbgOk;
+                if (NativeMethods.NtdbgOk != (ntdbgStatus =
+                        MemoryHelpers.WriteMemoryToFile(strArgFile, processId, startAddress, dataSize, fileMode)))
                 {
                     File.Delete(strArgFile);
 
                     Context.CONSOLE.Write("Couldn`t dump memory to file!");
-                    Context.CONSOLE.Write("Error code:" + ntdbgStatus.ToString() + " - " + NativeMethods.GetStatusString(ntdbgStatus) + ".");
+                    Context.CONSOLE.Write("Error code:" + ntdbgStatus + " - " +
+                                          NativeMethods.GetStatusString(ntdbgStatus) + ".");
                     return;
                 }
             }
             else
             {
                 bRet = MemoryHelpers.WriteMemoryToFile(strArgFile,
-                                                DkmMethods.GetDkmProcess(Context.IDE.Debugger.CurrentStackFrame),
-                                                startAddress,
-                                                dataSize,
-                                                fileMode);
+                    DkmMethods.GetDkmProcess(Context.IDE.Debugger.CurrentStackFrame),
+                    startAddress,
+                    dataSize,
+                    fileMode);
                 if (!bRet)
                 {
                     Context.CONSOLE.Write("Couldn`t dump memory to file!");
@@ -195,7 +186,7 @@ namespace VSDebugCoreLib.Commands.Memory
                 }
             }
 
-            Context.CONSOLE.Write("Wrote: " + dataSize.ToString() + " bytes to: " + MiscHelpers.GetClickableFileName(strArgFile));
+            Context.CONSOLE.Write("Wrote: " + dataSize + " bytes to: " + MiscHelpers.GetClickableFileName(strArgFile));
         }
     }
 }
